# tail: Performance improvements