Reasonable performance improvements for `tail` command. Benchmarks...
```
$ hyperfine -L tail /usr/bin/tail,./target/release/tail.original,./target/release/tail "{tail} -n 100000 ./shakespeare.txt"
Benchmark 1: /usr/bin/tail -n 100000 ./shakespeare.txt
  Time (mean ± σ):       4.8 ms ±   0.4 ms    [User: 2.6 ms, System: 2.3 ms]
  Range (min … max):     4.2 ms …   6.7 ms    539 runs
  
Benchmark 2: ./target/release/tail.original -n 100000 ./shakespeare.txt
  Time (mean ± σ):       4.6 ms ±   0.7 ms    [User: 3.1 ms, System: 1.5 ms]
  Range (min … max):     3.9 ms …  11.7 ms    466 runs
 
Benchmark 3: ./target/release/tail -n 100000 ./shakespeare.txt
  Time (mean ± σ):       3.6 ms ±   0.4 ms    [User: 2.2 ms, System: 1.4 ms]
  Range (min … max):     3.1 ms …   5.9 ms    689 runs
  
Summary
  ./target/release/tail -n 100000 ./shakespeare.txt ran
    1.28 ± 0.24 times faster than ./target/release/tail.original -n 100000 ./shakespeare.txt
    1.35 ± 0.18 times faster than /usr/bin/tail -n 100000 ./shakespeare.txt
```

```
$ hyperfine -L tail /usr/bin/tail,./target/release/tail.original,./target/release/tail "{tail} -n +100000 ./shakespeare.txt"
Benchmark 1: /usr/bin/tail -n +100000 ./shakespeare.txt
  Time (mean ± σ):       4.4 ms ±   0.3 ms    [User: 2.3 ms, System: 2.1 ms]
  Range (min … max):     3.9 ms …   5.8 ms    457 runs
  
Benchmark 2: ./target/release/tail.original -n +100000 ./shakespeare.txt
  Time (mean ± σ):       5.3 ms ±   0.4 ms    [User: 3.7 ms, System: 1.6 ms]
  Range (min … max):     4.6 ms …   8.1 ms    450 runs
 
Benchmark 3: ./target/release/tail -n +100000 ./shakespeare.txt
  Time (mean ± σ):       3.3 ms ±   0.6 ms    [User: 1.9 ms, System: 1.4 ms]
  Range (min … max):     2.7 ms …  10.1 ms    645 runs
 
Summary
  ./target/release/tail -n +100000 ./shakespeare.txt ran
    1.34 ± 0.26 times faster than /usr/bin/tail -n +100000 ./shakespeare.txt
    1.64 ± 0.32 times faster than ./target/release/tail.original -n +100000 ./shakespeare.txt
```